backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
redis>=4.5,<5.0
celery>=5.3,<6.0
gunicorn>=20.1,<21.0
openpyxl>=3.1,<4.0
pandas>=2.1,<3.0
whitenoise==6.8.2